Pages: Update schedule indicator to use relative time status component
We have a relative time status component that is used on scheduled post
lists to show when the post will be published (from the current time).
This commit updates the pages list to use the same component. When
viewing /pages/scheduled/, users should now see the indicator letting
them know when the page will be scheduled. When clicked the link will
either take them to the page editor or page preview when appropriate.

Fixes: #18418