DevDocs: Make component header links obvious
The DevDocs have a very useful feature where we can view a single component in isolation, rather than the full list. When developing an example, that feature massively improves the experience over constantly refreshing the full list and filtering.

Without some kind of indicator, though -- like a link icon or an underline -- it's easy to miss the fact that the title of each component is a link to the singular view of that component, leading to unnecessary frustration and wasted time.